// src/api/generated/models/menu-agent-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Enterprise API
 * React + TypeScript + Spring Boot OpenAPI
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface MenuAgentDTO {
    'id'?: number;
    'centerId'?: number;
    'upperId'?: number;
    'seq'?: number;
    'name'?: string;
    'path'?: string;
    'depth'?: number;
    'useFlag'?: number;
    'isLeaf'?: number;
}